+++ 0209-Minimum-Size-Subarray-Sum/cpp-0209/CMakeLists.txt
@@ -3,5 +3,5 @@ project(cpp_0209)
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
 
-set(SOURCE_FILES main3.cpp)
+set(SOURCE_FILES main4.cpp)
 add_executable(cpp_0209 ${SOURCE_FILES})
