+++ 0209-Minimum-Size-Subarray-Sum/cpp-0209/main3.cpp
@@ -5,46 +5,49 @@
 #include <iostream>
 #include <cassert>
 #include <vector>
+#include <algorithm>
 
 using namespace std;
 
-// Sliding Window
-// Time Complexity: O(n)
-// Space Complexity: O(1)
+
+// Sum Prefix + Binary Search
+// Time Complexity: O(nlogn)
+// Space Complexity: O(n)
 class Solution {
 public:
     int minSubArrayLen(int s, vector<int>& nums) {
 
         assert(s > 0);
 
-        int l = 0 , r = -1; // sliding window: nums[l...r]
-        int sum = 0;
+        vector<int> sums(nums.size() + 1, 0);
+        for(int i = 1 ; i <= nums.size() ; i ++)
+            sums[i] = sums[i-1] + nums[i-1];
+
         int res = nums.size() + 1;
-
-        while(l < nums.size()){
-
-            if(r + 1 < nums.size() && sum < s)
-                sum += nums[++r];
-            else
-                sum -= nums[l++];
-
-            if(sum >= s)
-                res = min(res, r - l + 1);
+        for(int l = 0 ; l < (int)nums.size() - 1 ; l ++){
+            auto r_bound = lower_bound(sums.begin(), sums.end(), sums[l] + s);
+            if(r_bound != sums.end()){
+                int r = r_bound - sums.begin();
+                res = min(res, r - l);
+            }
         }
 
-        if(res == nums.size() + 1)
-            return 0;
-        return res;
+        return res == nums.size() + 1 ? 0 : res;
     }
 };
 
 int main() {
 
-    int nums[] = {2, 3, 1, 2, 4, 3};
-    vector<int> vec(nums, nums + sizeof(nums)/sizeof(int));
-    int s = 7;
+    vector<int> nums1 = {2, 3, 1, 2, 4, 3};
+    int s1 = 7;
+    cout << Solution().minSubArrayLen(s1, nums1) << endl;
+    // 2
 
-    cout << Solution().minSubArrayLen(s, vec) << endl;
+    // ---
+
+    vector<int> nums2 = {};
+    int s2 = 100;
+    cout << Solution().minSubArrayLen(s2, nums2) << endl;
 
     return 0;
 }
